refactor(bench): extract assembleResult from the load-harness CLI
`cli.ts` reads `process.argv` and calls `main()` at import time, so
nothing declared in it — including the pure `RunResult` assembler and
its percentile helper — can be reached from a test. Moves the type and
the function verbatim into `assemble-result.ts`, and folds the `Variant`
union that `RunResult.run.variant` had been duplicating inline.
